backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
earthengine-api==1.1.2
pandas==2.2.2
prophet==1.1.6
pydantic==2.9.2
joblib==1.4.2
numpy==1.26.4
python-dotenv==1.0.1
requests==2.32.3
sentence-transformers==3.0.1

